requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pypdf
chromadb
streamlit
fastapi
uvicorn
tiktoken
python-dotenv